Rearranging getters and setters to utils, removing unused code.
